tidy domain encoder
